+++ tests/test_render.py
@@ -67,6 +67,30 @@ class RenderTest(TestCase):
         for point, ts in data:
             self.assertEqual(point, 12)
 
+    def test_correct_timezone(self):
+        response = self.app.get(self.url, query_string={
+            'target': 'constantLine(12)',
+            'format': 'json',
+            'from': '07:00_20140226',
+            'until': '08:00_20140226',
+            # tz is UTC
+        })
+        data = json.loads(response.data.decode('utf-8'))[0]['datapoints']
+
+        # all the from/until/tz combinations lead to the same window
+        expected = [[12, 1393398000], [12, 1393401600]]
+        self.assertEqual(data, expected)
+
+        response = self.app.get(self.url, query_string={
+            'target': 'constantLine(12)',
+            'format': 'json',
+            'from': '08:00_20140226',
+            'until': '09:00_20140226',
+            'tz': 'Europe/Berlin',
+        })
+        data = json.loads(response.data.decode('utf-8'))[0]['datapoints']
+        self.assertEqual(data, expected)
+
     def test_render_options(self):
         self.create_db()
         db2 = os.path.join(WHISPER_DIR, 'foo.wsp')